backend/requirements.txt
djangorestframework
mysqlclient
web3
pillow
qrcode